pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neo-triage"
version = "0.0.1"
description = "Bayesian NEO follow-up prioritization with Claude 4.7 as the astronomy reasoning engine."
authors = [
    { name = "Kacper Saks" },
    { name = "Paweł Kulak" },
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "anthropic>=0.25",
    "polars>=0.20",
    "scikit-learn>=1.4",
    "websockets>=12.0",
    "httpx>=0.27",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.9",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "RET", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
# SYSTEM_PROMPT is prose; readability of the prompt text matters more than 100-char wrapping.
"backend/services/briefing_engine.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
